+++ src/realm/table.cpp
@@ -3954,6 +3954,15 @@ size_t Table::find_first(size_t col_ndx, util::Optional<double> value) const
 {
     return value ? find_first(col_ndx, *value) : find_first_null(col_ndx);
 }
+
+// Explicitly instantiate the generic case of the template for the types we care about.
+template size_t Table::find_first(size_t col_ndx, bool) const;
+template size_t Table::find_first(size_t col_ndx, int64_t) const;
+template size_t Table::find_first(size_t col_ndx, float) const;
+template size_t Table::find_first(size_t col_ndx, double) const;
+template size_t Table::find_first(size_t col_ndx, util::Optional<bool>) const;
+template size_t Table::find_first(size_t col_ndx, util::Optional<int64_t>) const;
+
 } // namespace realm
 
 size_t Table::find_first_link(size_t target_row_index) const